ROP_project/requirements.txt
ultralytics
opencv-python
matplotlib
tqdm
torch
torchvision
pyyaml
numpy
pandas
scikit-learn
